requirements.txt
openai
langchain
streamlit
google.generativeai
python-dotenv
langchain-google-vertexai
langchain_google_genai
pathlib
textwrap3
IPython
 